load mode from stream

+++ src/extra/geometry_viewer/models/autd3_model.cpp
@@ -3,7 +3,7 @@
 // Created Date: 18/10/2022
 // Author: Shun Suzuki
 // -----
-// Last Modified: 03/11/2022
+// Last Modified: 16/03/2023
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -14,4 +14,4 @@
 uint8_t model_data[] = {
 #include "AUTD3.glb.txt"
 };
-unsigned model_size = sizeof model_data;
+size_t model_size = sizeof model_data;
